modules/poc-api/src/requirements.txt
fastapi
uvicorn
boto3
python-multipart
pydantic
psycopg2-binary
sqlalchemy>=2.0.0
python-dotenv
requests
pytest
httpx